warning Fix
Bench: 2084468

+++ src/misc.h
@@ -1181,11 +1181,10 @@ inline std::uint64_t hash_string(std::string_view str) noexcept {
 
 template<typename T>
 std::uint64_t hash_raw_data(const T& value) noexcept {
-    // Must have no padding bytes because reinterpreting as char
+    // Must have no padding bytes because reinterpreting as char*
     static_assert(std::has_unique_object_representations<T>());
 
-    return static_cast<std::size_t>(
-      hash_bytes(reinterpret_cast<const char*>(&value), sizeof(value)));
+    return hash_bytes(reinterpret_cast<const char*>(&value), sizeof(value));
 }
 
 template<typename T>
